refactor modules (#79)
