+++ apps/api-journeys/src/application/application.ts
@@ -1,0 +1,36 @@
+import { createApplication } from 'graphql-modules'
+import { schemaBuilder } from '@core/shared/util-graphql'
+import {
+  actionModule,
+  blockModule,
+  buttonModule,
+  cardModule,
+  iconModule,
+  imageModule,
+  journeyModule,
+  radioQuestionModule,
+  responseModule,
+  signUpModule,
+  stepModule,
+  typographyModule,
+  videoModule
+} from '../modules'
+
+export const application = createApplication({
+  modules: [
+    actionModule,
+    blockModule,
+    buttonModule,
+    cardModule,
+    iconModule,
+    imageModule,
+    journeyModule,
+    radioQuestionModule,
+    responseModule,
+    signUpModule,
+    stepModule,
+    typographyModule,
+    videoModule
+  ],
+  schemaBuilder
+})
