+++ apps/api-journeys/src/application/index.ts
@@ -1,0 +1,1 @@
+export { application } from './application'
