+++ apps/api-journeys/src/modules/block/transformBlock/index.ts
@@ -1,2 +1,1 @@
 export { transformBlock } from './transformBlock'
-export { blockModule } from './block'
