+++ apps/api-journeys/src/modules/button/index.ts
@@ -1,0 +1,1 @@
+export { buttonModule } from './button'
