+++ apps/api-journeys/src/modules/card/card.spec.ts
@@ -1,0 +1,88 @@
+import { testkit, gql } from 'graphql-modules'
+import { schemaBuilder } from '@core/shared/util-graphql'
+import { cardModule, journeyModule, blockModule } from '..'
+import dbMock from '../../../tests/dbMock'
+import { v4 as uuidv4 } from 'uuid'
+import { Block, ThemeName, ThemeMode } from '.prisma/api-journeys-client'
+import { DocumentNode, ExecutionResult } from 'graphql'
+
+describe('CardModule', () => {
+  let app, journeyId
+
+  beforeEach(() => {
+    app = testkit.testModule(cardModule, {
+      schemaBuilder,
+      modules: [journeyModule, blockModule]
+    })
+    journeyId = uuidv4()
+    dbMock.journey.findUnique.mockResolvedValue({
+      id: journeyId,
+      title: 'published',
+      published: true,
+      locale: 'en-US',
+      themeMode: ThemeMode.light,
+      themeName: ThemeName.base
+    })
+  })
+
+  async function query(document: DocumentNode): Promise<ExecutionResult> {
+    return await testkit.execute(app, {
+      document,
+      variableValues: {
+        id: journeyId
+      },
+      contextValue: {
+        db: dbMock
+      }
+    })
+  }
+
+  describe('CardBlock', () => {
+    it('returns CardBlock', async () => {
+      const parentBlockId = uuidv4()
+      const coverBlockId = uuidv4()
+      const card: Block = {
+        id: uuidv4(),
+        journeyId,
+        blockType: 'CardBlock',
+        parentBlockId,
+        parentOrder: 0,
+        extraAttrs: {
+          backgroundColor: '#FFF',
+          coverBlockId,
+          themeMode: ThemeMode.light,
+          themeName: ThemeName.base
+        }
+      }
+      dbMock.block.findMany.mockResolvedValue([card])
+      const { data } = await query(gql`
+        query ($id: ID!) {
+          journey(id: $id) {
+            blocks {
+              id
+              __typename
+              parentBlockId
+              ... on CardBlock {
+                backgroundColor
+                coverBlockId
+                themeMode
+                themeName
+              }
+            }
+          }
+        }
+      `)
+      expect(data?.journey.blocks).toEqual([
+        {
+          id: card.id,
+          __typename: 'CardBlock',
+          parentBlockId,
+          backgroundColor: '#FFF',
+          coverBlockId,
+          themeMode: ThemeMode.light,
+          themeName: ThemeName.base
+        }
+      ])
+    })
+  })
+})
