+++ apps/api-journeys/src/modules/card/index.ts
@@ -1,0 +1,1 @@
+export { cardModule } from './card'
