+++ apps/api-journeys/src/modules/image/index.ts
@@ -1,0 +1,1 @@
+export { imageModule } from './image'
