+++ apps/api-journeys/src/modules/radioQuestion/index.ts
@@ -1,0 +1,1 @@
+export { radioQuestionModule } from './radioQuestion'
