+++ apps/api-journeys/src/modules/step/index.ts
@@ -1,0 +1,1 @@
+export { stepModule } from './step'
