+++ apps/api-journeys/src/modules/typography/index.ts
@@ -1,0 +1,1 @@
+export { typographyModule } from './typography'
