+++ apps/api-journeys/src/modules/video/index.ts
@@ -1,0 +1,1 @@
+export { videoModule } from './video'
